pyproject.toml
[tool.poetry]
name = "ghost_post_rest"
version = "0.1.0"
description = ""
authors = ["Andrew Belanger <andrew.j.belanger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^2.2.9"
django-cors-headers = "^3.5.0"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
